+++ extra-projects/AdventureWorks-EF9/Program.cs
@@ -1,0 +1,10 @@
+namespace AdventureWorks_EF9
+{
+    internal class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hello, World!");
+        }
+    }
+}