+++ src/Wolfgang.DbContextBuilder-Core/SqliteDbContextCreator.cs
@@ -29,12 +29,8 @@ internal sealed class SqliteDbContextCreator : ICreateDbContext, IDisposable
     /// <inheritdoc />
     public Task<TDbContext> CreateDbContextAsync<TDbContext>(DbContextOptionsBuilder<TDbContext> optionsBuilder) where TDbContext : DbContext
     {
-
         optionsBuilder.UseSqlite(_connection);
-
-        var options = optionsBuilder.Options;
-
-        return Task.FromResult((TDbContext)Activator.CreateInstance(typeof(TDbContext), options)!);
+        return Task.FromResult(DbContextActivator<TDbContext>.Create(optionsBuilder.Options));
     }